+++ rc.rkt
@@ -100,6 +100,7 @@
 (define-xiden-setting XIDEN_TRUST_ANY_PUBLIC_KEY boolean? #f)
 (define-xiden-setting XIDEN_TRUST_BAD_DIGEST boolean? #f)
 (define-xiden-setting XIDEN_TRUST_BAD_SIGNATURE boolean? #f)
+(define-xiden-setting XIDEN_TRUST_CERTIFICATES (listof path-string?) null)
 (define-xiden-setting XIDEN_TRUST_UNSIGNED boolean? #f)
 (define-xiden-setting XIDEN_TRUST_UNVERIFIED_HOST boolean? #f)
 (define-xiden-setting XIDEN_VERBOSE boolean? #f)